+++ src/unix/apple/app_store/mod.rs
@@ -1,4 +1,5 @@
 // Take a look at the license at the top of the repository in the LICENSE file.
 
 pub mod component;
+#[cfg(feature = "system")]
 pub mod process;
