+++ src/unix/apple/macos/process.rs
@@ -110,7 +110,7 @@ impl ProcessInner {
     }
 
     pub(crate) fn kill_with(&self, signal: Signal) -> Option<bool> {
-        let c_signal = crate::sys::convert_signal(signal)?;
+        let c_signal = crate::sys::system::convert_signal(signal)?;
         unsafe { Some(kill(self.pid.0, c_signal) == 0) }
     }
