+++ src/unix/freebsd/process.rs
@@ -69,7 +69,7 @@ pub(crate) struct ProcessInner {
 
 impl ProcessInner {
     pub(crate) fn kill_with(&self, signal: Signal) -> Option<bool> {
-        let c_signal = crate::sys::convert_signal(signal)?;
+        let c_signal = crate::sys::system::convert_signal(signal)?;
         unsafe { Some(libc::kill(self.pid.0, c_signal) == 0) }
     }
