+++ src/windows/system.rs
@@ -4,32 +4,43 @@ use crate::{Cpu, CpuRefreshKind, LoadAvg, MemoryRefreshKind, Pid, ProcessRefresh
 
 use crate::sys::cpu::*;
 use crate::sys::process::get_start_time;
-use crate::sys::tools::*;
-use crate::sys::utils::{get_now, get_reg_string_value, get_reg_value_u32};
 use crate::{Process, ProcessInner};
 
 use crate::utils::into_iter;
 
 use std::cell::UnsafeCell;
 use std::collections::HashMap;
-use std::ffi::OsString;
+use std::ffi::{OsStr, OsString};
 use std::mem::{size_of, zeroed};
-use std::os::windows::ffi::OsStringExt;
+use std::os::windows::ffi::{OsStrExt, OsStringExt};
 use std::ptr;
-use std::time::SystemTime;
+use std::time::{Duration, SystemTime};
 
 use ntapi::ntexapi::SYSTEM_PROCESS_INFORMATION;
-use windows::core::PWSTR;
+use windows::core::{PCWSTR, PWSTR};
 use windows::Wdk::System::SystemInformation::{NtQuerySystemInformation, SystemProcessInformation};
-use windows::Win32::Foundation::{HANDLE, STATUS_INFO_LENGTH_MISMATCH, STILL_ACTIVE};
+use windows::Win32::Foundation::{self, HANDLE, STATUS_INFO_LENGTH_MISMATCH, STILL_ACTIVE};
 use windows::Win32::System::ProcessStatus::{K32GetPerformanceInfo, PERFORMANCE_INFORMATION};
-use windows::Win32::System::Registry::HKEY_LOCAL_MACHINE;
+use windows::Win32::System::Registry::{
+    RegCloseKey, RegOpenKeyExW, RegQueryValueExW, HKEY, HKEY_LOCAL_MACHINE, KEY_READ, REG_NONE,
+};
 use windows::Win32::System::SystemInformation;
 use windows::Win32::System::SystemInformation::{
     ComputerNamePhysicalDnsHostname, GetComputerNameExW, GetTickCount64, GlobalMemoryStatusEx,
     MEMORYSTATUSEX, SYSTEM_INFO,
 };
 use windows::Win32::System::Threading::GetExitCodeProcess;
+
+declare_signals! {
+    (),
+    Signal::Kill => (),
+    _ => None,
+}
+
+#[doc = include_str!("../../md_doc/supported_signals.md")]
+pub const SUPPORTED_SIGNALS: &[crate::Signal] = supported_signals();
+#[doc = include_str!("../../md_doc/minimum_cpu_update_interval.md")]
+pub const MINIMUM_CPU_UPDATE_INTERVAL: Duration = Duration::from_millis(200);
 
 const WINDOWS_ELEVEN_BUILD_NUMBER: u32 = 22000;
 
@@ -593,4 +604,124 @@ fn get_dns_hostname() -> Option<String> {
 
     sysinfo_debug!("Failed to get computer hostname");
     None
+}
+
+fn add_english_counter(
+    s: String,
+    query: &mut super::cpu::Query,
+    keys: &mut Option<KeyHandler>,
+    counter_name: String,
+) {
+    let mut full = s.encode_utf16().collect::<Vec<_>>();
+    full.push(0);
+    if query.add_english_counter(&counter_name, full) {
+        *keys = Some(KeyHandler::new(counter_name));
+    }
+}
+
+fn get_now() -> u64 {
+    SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .map(|n| n.as_secs())
+        .unwrap_or(0)
+}
+
+fn utf16_str<S: AsRef<OsStr> + ?Sized>(text: &S) -> Vec<u16> {
+    OsStr::new(text)
+        .encode_wide()
+        .chain(Some(0))
+        .collect::<Vec<_>>()
+}
+
+struct RegKey(HKEY);
+
+impl RegKey {
+    unsafe fn open(hkey: HKEY, path: &[u16]) -> Option<Self> {
+        let mut new_hkey = Default::default();
+        if RegOpenKeyExW(
+            hkey,
+            PCWSTR::from_raw(path.as_ptr()),
+            0,
+            KEY_READ,
+            &mut new_hkey,
+        )
+        .is_err()
+        {
+            return None;
+        }
+        Some(Self(new_hkey))
+    }
+
+    unsafe fn get_value(
+        &self,
+        field_name: &[u16],
+        buf: &mut [u8],
+        buf_len: &mut u32,
+    ) -> windows::core::Result<()> {
+        let mut buf_type = REG_NONE;
+
+        RegQueryValueExW(
+            self.0,
+            PCWSTR::from_raw(field_name.as_ptr()),
+            None,
+            Some(&mut buf_type),
+            Some(buf.as_mut_ptr()),
+            Some(buf_len),
+        )
+        .ok()
+    }
+}
+
+impl Drop for RegKey {
+    fn drop(&mut self) {
+        let _err = unsafe { RegCloseKey(self.0) };
+    }
+}
+
+pub(crate) fn get_reg_string_value(hkey: HKEY, path: &str, field_name: &str) -> Option<String> {
+    let c_path = utf16_str(path);
+    let c_field_name = utf16_str(field_name);
+
+    unsafe {
+        let new_key = RegKey::open(hkey, &c_path)?;
+        let mut buf_len: u32 = 2048;
+        let mut buf: Vec<u8> = Vec::with_capacity(buf_len as usize);
+
+        loop {
+            match new_key.get_value(&c_field_name, &mut buf, &mut buf_len) {
+                Ok(()) => break,
+                Err(err) if err.code() == Foundation::ERROR_MORE_DATA.to_hresult() => {
+                    // Needs to be updated for `Vec::reserve` to actually add additional capacity.
+                    buf.set_len(buf.capacity());
+                    buf.reserve(buf_len as _);
+                }
+                _ => return None,
+            }
+        }
+
+        buf.set_len(buf_len as _);
+
+        let words = std::slice::from_raw_parts(buf.as_ptr() as *const u16, buf.len() / 2);
+        let mut s = String::from_utf16_lossy(words);
+        while s.ends_with('\u{0}') {
+            s.pop();
+        }
+        Some(s)
+    }
+}
+
+pub(crate) fn get_reg_value_u32(hkey: HKEY, path: &str, field_name: &str) -> Option<[u8; 4]> {
+    let c_path = utf16_str(path);
+    let c_field_name = utf16_str(field_name);
+
+    unsafe {
+        let new_key = RegKey::open(hkey, &c_path)?;
+        let mut buf_len: u32 = 4;
+        let mut buf = [0u8; 4];
+
+        new_key
+            .get_value(&c_field_name, &mut buf, &mut buf_len)
+            .map(|_| buf)
+            .ok()
+    }
 }
